Improve gun rotation
The gun was calculating its direction based on the position of the tank, not its own position. By calculating the angle between the gun and mouse it is slightly more accurate.

The tanks relative rotation was being set instead of its global rotation. This meant that we had to subtract the rotation of the tank to get the correct value. By setting the global rotation of the gun instead we can ignore the rotation of the tank.

Indented comments.